Grid Manager Updated
Functions for Oddr (Unity coords) and Cube (easy algorithm coords) work up through the *Range functions. Those return a list of Vector3Int's that are within int d tiles of a target Vector3Int. This will be the basis for updating tiles when a water tile is pooled or soaked, and may be useful for recipe cluster finding.