Merge Tile
Progress on merging tiles. TODOs are commented in the code where work needs to be done. Biggest thing is finding a way to import a vector math package of some kind so we can do vector math. Also, needs a pair of dictionaries: one that stores a plant name and its unique vector, and a dictionary that stores a recipe result and its cost vector.